# Precompile regexps before using them in a loop.