Precompiling some of the regular expressions significantly speeds up initial loading of the database, `search` and `search_stresses`.

From a simple benchmark:

Function        | original time | optimized time | speedup
----------------|---------------|----------------|----------
init_cmu        | 0.5609 s      | 0.3812 s       | 1.47x
search          | 0.1849 s      | 0.0494 s       | 3.74x
search_stresses | 0.7285 s      | 0.6404 s       | 1.13x