@aparrish 2015-06-23
this is awesome! "low hanging optimization fruit" was next on my to-do list and you just took care of a big chunk of it for me. :) thanks!